Update handleError to new axios interceptor
